import lib to main, bin print, less redundant dir
Begin to use CaniPy code in the gui script, for now just bare control of the serial object within the CaniPy instance. More efficient hex printing implemented. Using "utils" naming instead of "canipy" to make it less redundant. Package script is still named "canipy" though. When imported by an external script if this is a module/package aliased to a project root, it is now set to allow it via "from canipy import CaniPy".