util improvements
Several updates to the utility code, including functions and additional lambdas for things like resetting port and baud assignments, current channel information instead of specifying a channel number, channel var for tracking, and safety code for whenever a serial device is not ready or channel number is not in 2-byte range. Override function implemented for debugging purposes.