better rx thread handling, fstrings
The GUI's COM thread is now determined by whether the window is open or not, rather than requiring user intervention with a button to manually halt it. It now halts when the window is closed, allowing for graceful exit. Log text now uses f-strings and opening COM reports the port and baud rate. Might have to make those settings prettier later.