mv'd debug to term, toolbar era
All debug output for the GUI now goes to the command line. The window now acts as a sort of toolbar with a set of buttons to set up and send commands to the receiver. It's recommended to use this "toolbar" in conjunction with the command output to determine activity.