move funcs to module, u8 parse, verbose handling
This is the first build desginated as "alpha" quality. Along with some updated return code support, some printout and support functions have been moved over to the module. Changed to using UTF-8 for hex parsing instead of ASCII. Mitigated most of the printout by storing them under a "verbose" flag, and will only appear if verbose is set to true in the canipy instance.