watch signal feature, evt driven sig rx, rm wx debug & ch status btns
Second alpha release. Added signal status monitor handling as an event-driven RX test. Click on the "Watch Signal" button to enable signal status monitoring on the radio. The console will log changes or updates in signal reception as reported by the radio. WX debug and channel status buttons have been removed as these are mainly for checking if WX receivers specifically are working and ch status can be confusing to use as this does tune you out of the current channel. These may be implemented in some form later in the future, but may still be accessed when importing the module programmatically.