purged most of pcrtest pcap doc
Removed a majority of the logged material in "pcrtest.txt". Pretty much all commands that were removed are covered already. About the only things left are the firmware version fetching 70 05, and implementing channel monitoring but that's in the pipeline already. Kept certain things for reference, but eventually this file will no longer be needed.