handling malformed act status, gui pwr off no pwr sav flag
E2 and 06 0B indicate when activation status loaded by the receiver is malformed or corrupted, necessitating a cold restart or in rare circumstances a refresh. Powering off the radio in GUI now does not specify power save as it might be causing unintended side effects, though further testing will be needed to figure out the upsides of using power saving in power off commands.